main 페이지 시작.

+++ ShinKimChi/src/main/java/org/skc/page/PageMaker.java
@@ -44,11 +44,11 @@ public class PageMaker {
 		double perPageNum = cri.getPerPageNum();
 		double displayPageNum = cri.getDisplayPageNum();
 		
-		double tempLast = Math.ceil(page/(double)displayPageNum) * displayPageNum;
+		double tempLast = Math.ceil(page/displayPageNum) * displayPageNum;
 		
 		first = (int)(tempLast - displayPageNum) + 1;
 		
-		last = (int)(totalCount < tempLast * perPageNum ? Math.ceil(totalCount/(double)perPageNum):tempLast);
+		last = (int)(totalCount < tempLast * perPageNum ? Math.ceil(totalCount/perPageNum):tempLast);
 		
 		prev = first > 1 ? first - 1 : 0;
 		next = totalCount > last * perPageNum ? last + 1 : 0;
